% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcs.R
\name{ds.kcs}
\alias{ds.kcs}
\title{List the KCs available within a partiuclar KC model of a DataShop student-step export}
\usage{
ds.kcs(stu.step, model)
}
\arguments{
\item{stu.step}{A DataShop Student-Step export data.frame}

\item{model}{The name of a KC model within the dataset}
}
\description{
List the KCs available within a partiuclar KC model of a DataShop student-step export
}

