% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ds.models}
\alias{ds.models}
\title{List the available KC models in a Student-Step file.}
\usage{
ds.models(stu.step)
}
\arguments{
\item{stu.step}{A Student-Step export from DataShop}
}
\description{
List the available KC models in a Student-Step file.
}

