% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ds.plot}
\alias{ds.plot}
\title{A Smooth Plot DataShop Learning Curves}
\usage{
ds.plot(stu.step, model, kc = "all", line.type = "smooth", title = NULL,
  legend.position = "none", label.scale = 1)
}
\arguments{
\item{stu.step}{A data.frame of a direct load of a student-step rollup file.}

\item{model}{The name of a KC model to use as it appears in DataShop}

\item{kc}{The name of a particular Kc to get individual kc curves or its number in the list of kcs}

\item{line.type}{Controls whether to render a "smooth" loess line or "discrete" true averages line.}

\item{title}{A title for the plot}

\item{legend.position}{defines the position of a legend, defaults to "none" because I usually explain the legend in a figure capion.}

\item{label.scale}{A scaling factor for the axis label and title text.}
}
\description{
Creates a loess smoothed ggplot from a student step roll up output from Datashop.
}
\keyword{ds.plot}

