% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_read.R
\name{ds.read}
\alias{ds.read}
\title{Read in a DataShop Student-Step rollup file}
\usage{
ds.read(file.name)
}
\arguments{
\item{file.name}{The name of the file to read in.}
}
\description{
Read in a DataShop Student-Step rollup file
}

