% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ds.save}
\alias{ds.save}
\title{Save a ds.plot in a commonly useful png format.}
\usage{
ds.save(plot, name, width = 3.3, height = 2.08, dpi = 300)
}
\arguments{
\item{plot}{The plot data}

\item{name}{The desired filename}

\item{width}{The width in inches (defaults to 3, which is recommended width for AcM/EDM format)}

\item{height}{The height in inches (which I default to 2.5)}

\item{dpi}{The dpi to output at (defaults to 300)}
}
\description{
This is just a basic wrapper around ggplot's ggsave but the default settings make it perfect for fitting into 
a single column of the ACM/EDM paper format.
}

